peristaltic-pump recipe: make `inputs` a Table, not an Array
`[inputs]` is now a TOML Table (a key/value store).  Each input is
represented in the `[inputs]` table by a key/value pair.

The key is a unique name (which also maps to the recipe file for building
that input thing).

The value is another table describing that input, that we'll represent
with an Input struct.